requirements.txt
streamlit==1.54.0
google-genai==1.65.0
gspread==6.2.1
pandas==2.3.3
